Highlight playlist name when selected
