+++ src/styles/main.scss
@@ -48,11 +48,12 @@ section.sidebar {
     height: 100%;
     padding: 20px;
     overflow-y: auto;
-    // background-color: red;
 
-    // .sidebar-item {
-    //     height: 100%;
-    // }
+
+    .selected {
+        color: var(--primary-color);
+        font-weight: 600;
+    }
 }
 
 section.main {
